test(llc): migrate client with-persistence tests to stream_chat_test
