+++ packages/stream_chat/test/src/client/client_test.dart
@@ -483,712 +483,6 @@ void main() {
     );
   });
 
-  group('Client with connected user with persistence', () {
-    const apiKey = 'test-api-key';
-    late final api = FakeChatApi();
-    late final persistence = MockPersistenceClient();
-
-    final user = User(id: 'test-user-id');
-    final token = Token.development(user.id).rawValue;
-
-    late StreamChatClient client;
-
-    setUpAll(() {
-      // fallback values
-      registerFallbackValue(FakeEvent());
-      registerFallbackValue(const PaginationParams());
-      registerFallbackValue(FakeChannelState());
-      registerFallbackValue(const Filter.empty());
-    });
-
-    setUp(() async {
-      when(() => persistence.updateLastSyncAt(any())).thenAnswer((_) => Future.value());
-      when(persistence.getLastSyncAt).thenAnswer((_) async => null);
-      final ws = FakeWebSocket();
-      client = StreamChatClient(apiKey, chatApi: api, ws: ws)..chatPersistenceClient = persistence;
-      await client.connectUser(user, token);
-      await delay(300);
-      expect(client.persistenceEnabled, isTrue);
-      expect(client.wsConnectionStatus, ConnectionStatus.connected);
-    });
-
-    tearDown(() async {
-      await client.dispose();
-    });
-
-    group('`.sync`', () {
-      test(
-        '''should update persistence connectionInfo and lastSync when sync succeeds''',
-        () async {
-          // persistence.updateLastSyncAt might be called
-          // when connecting the user.
-          // Resetting the logs so we start counting invocations correctly.
-          reset(persistence);
-          const cids = ['test-cid-1', 'test-cid-2', 'test-cid-3'];
-          final lastSyncAt = DateTime.now();
-          when(() => api.general.sync(cids, lastSyncAt)).thenAnswer(
-            (_) async => SyncResponse()
-              ..events = [
-                Event(
-                  isLocal: false,
-                  type: EventType.healthCheck,
-                  connectionId: 'test-connection-id',
-                  me: OwnUser.fromUser(user),
-                ),
-                Event(
-                  isLocal: false,
-                  type: EventType.messageDeleted,
-                  message: Message(id: 'test-message-id'),
-                ),
-              ],
-          );
-
-          when(() => persistence.updateConnectionInfo(any())).thenAnswer((_) => Future.value());
-          when(() => persistence.updateLastSyncAt(any())).thenAnswer((_) => Future.value());
-
-          await client.sync(cids: cids, lastSyncAt: lastSyncAt);
-
-          verify(() => persistence.updateConnectionInfo(any())).called(1);
-          verify(() => persistence.updateLastSyncAt(any())).called(1);
-          verify(() => api.general.sync(cids, lastSyncAt)).called(1);
-        },
-      );
-
-      test(
-        'should work fine if persistence contains sync params',
-        () async {
-          // persistence.updateLastSyncAt might be called
-          // when connecting the user.
-          // Resetting the logs so we start counting invocations correctly.
-          reset(persistence);
-          const cids = ['test-cid-1', 'test-cid-2', 'test-cid-3'];
-          final lastSyncAt = DateTime.now();
-
-          when(persistence.getChannelCids).thenAnswer((_) async => cids);
-          when(persistence.getLastSyncAt).thenAnswer((_) async => lastSyncAt);
-
-          when(() => api.general.sync(cids, lastSyncAt)).thenAnswer(
-            (_) async => SyncResponse()
-              ..events = [
-                Event(
-                  isLocal: false,
-                  type: EventType.healthCheck,
-                  connectionId: 'test-connection-id',
-                  me: OwnUser.fromUser(user),
-                ),
-                Event(
-                  isLocal: false,
-                  type: EventType.messageDeleted,
-                  message: Message(id: 'test-message-id', text: 'Hey!'),
-                ),
-              ],
-          );
-
-          when(() => persistence.updateConnectionInfo(any())).thenAnswer((_) => Future.value());
-          when(() => persistence.updateLastSyncAt(any())).thenAnswer((_) => Future.value());
-
-          await client.sync();
-
-          verify(() => persistence.updateConnectionInfo(any())).called(1);
-          verify(() => persistence.updateLastSyncAt(any())).called(1);
-          verify(() => api.general.sync(cids, lastSyncAt)).called(1);
-          verify(persistence.getChannelCids).called(1);
-          verify(persistence.getLastSyncAt).called(1);
-        },
-      );
-    });
-
-    group('`.queryChannels`', () {
-      test(
-        'should emit channels twice if persistence contains some channels',
-        () async {
-          final persistentChannelStates = List.generate(
-            3,
-            (index) => ChannelState(
-              channel: ChannelModel(cid: 'test-type-$index:test-id-$index'),
-            ),
-          );
-
-          when(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer((_) async => QueryChannelsResponse()..channels = persistentChannelStates);
-
-          final channelStates = List.generate(
-            3,
-            (index) => ChannelState(
-              channel: ChannelModel(cid: 'test-type-$index:test-id-$index'),
-            ),
-          );
-
-          when(
-            () => api.channel.queryChannels(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer(
-            (_) async => QueryChannelsResponse()..channels = channelStates,
-          );
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer(
-            (_) async => <String, List<Message>>{
-              for (final channelState in channelStates)
-                channelState.channel!.cid: [Message(id: 'test-message-id', text: 'Test message')],
-            },
-          );
-
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async {});
-          when(
-            () => persistence.saveChannelQueries(
-              cids: any(named: 'cids'),
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              resolvedFilter: any(named: 'resolvedFilter'),
-              resolvedSort: any(named: 'resolvedSort'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              clearQueryCache: any(named: 'clearQueryCache'),
-            ),
-          ).thenAnswer((_) => Future.value());
-
-          // setUp's `connectUser` schedules debounced persistence writes
-          // (1s window) that would otherwise fire during this test's wait
-          // and pollute the call counts. Wait past the debounce, then clear.
-          await delay(1100);
-          clearInteractions(persistence);
-
-          await expectLater(
-            client.queryChannels(),
-            emitsInOrder([
-              // emits persistent channels first
-              persistentChannelStates.map(isCorrectChannelFor),
-              // makes api call and emits network fetched channels
-              channelStates.map(isCorrectChannelFor),
-            ]),
-          );
-
-          // Wait safely past the 1s debounce on persistence writes
-          // (updateChannelState, updateChannelThreads) so all trailing
-          // invocations have fired before we verify counts.
-          await delay(1500);
-
-          verify(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).called(1);
-
-          verify(
-            () => api.channel.queryChannels(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).called(1);
-
-          verify(() => persistence.getChannelThreads(any())).called(channelStates.length);
-          verify(() => persistence.updateChannelState(any())).called(channelStates.length);
-          verify(() => persistence.updateChannelThreads(any(), any())).called(channelStates.length);
-          verify(
-            () => persistence.saveChannelQueries(
-              cids: any(named: 'cids'),
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              resolvedFilter: any(named: 'resolvedFilter'),
-              resolvedSort: any(named: 'resolvedSort'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              clearQueryCache: any(named: 'clearQueryCache'),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        '''should never rethrow network call if persistence already emitted some channels''',
-        () async {
-          final persistentChannelStates = List.generate(
-            3,
-            (index) => ChannelState(
-              channel: ChannelModel(cid: 'test-type-$index:test-id-$index'),
-            ),
-          );
-
-          when(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer((_) async => QueryChannelsResponse()..channels = persistentChannelStates);
-
-          when(
-            () => api.channel.queryChannels(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenThrow(StreamChatNetworkError(ChatErrorCode.inputError));
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer(
-            (_) async => <String, List<Message>>{
-              for (final channelState in persistentChannelStates)
-                channelState.channel!.cid: [Message(id: 'test-message-id', text: 'Test message')],
-            },
-          );
-
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async => {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async => {});
-
-          // setUp's `connectUser` schedules debounced persistence writes
-          // (1s window) that would otherwise fire during this test's wait
-          // and pollute the call counts. Wait past the debounce, then clear.
-          await delay(1100);
-          clearInteractions(persistence);
-
-          await expectLater(
-            client.queryChannels(),
-            emitsInOrder([
-              // emits persistent channels
-              persistentChannelStates.map(isCorrectChannelFor),
-            ]),
-          );
-
-          // Wait safely past the 1s debounce on persistence writes
-          // (updateChannelState, updateChannelThreads) so all trailing
-          // invocations have fired before we verify counts.
-          await delay(1500);
-
-          verify(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).called(1);
-
-          verify(
-            () => api.channel.queryChannels(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).called(1);
-
-          verify(() => persistence.getChannelThreads(any())).called(persistentChannelStates.length);
-          verify(() => persistence.updateChannelState(any())).called(persistentChannelStates.length);
-          verify(() => persistence.updateChannelThreads(any(), any())).called(persistentChannelStates.length);
-        },
-      );
-
-      test(
-        'queryChannelsOnline with inline filter persists via saveChannelQueries',
-        () async {
-          final filter = Filter.in_('members', const ['test-user-id']);
-
-          final channelStates = List.generate(
-            3,
-            (i) => ChannelState(channel: ChannelModel(cid: 'test-type-$i:test-id-$i')),
-          );
-
-          when(
-            () => api.channel.queryChannels(
-              filter: filter,
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer(
-            (_) async => QueryChannelsResponse()..channels = channelStates,
-          );
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer((_) async => <String, List<Message>>{});
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async {});
-          when(
-            () => persistence.saveChannelQueries(
-              cids: any(named: 'cids'),
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              resolvedFilter: any(named: 'resolvedFilter'),
-              resolvedSort: any(named: 'resolvedSort'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              clearQueryCache: any(named: 'clearQueryCache'),
-            ),
-          ).thenAnswer((_) => Future.value());
-
-          await delay(1100);
-          clearInteractions(persistence);
-
-          await client.queryChannelsOnline(filter: filter);
-
-          // The standard path passes filter (the inline filter) and a null
-          // predefinedFilter. resolvedFilter / resolvedSort stay null —
-          // they're only meaningful for the predefined-filter path.
-          verify(
-            () => persistence.saveChannelQueries(
-              cids: channelStates.map((s) => s.channel!.cid).toList(),
-              filter: filter,
-              sort: null,
-              predefinedFilter: null,
-              resolvedFilter: null,
-              resolvedSort: null,
-              filterValues: null,
-              sortValues: null,
-              clearQueryCache: true,
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'queryChannelsOnline with predefined filter persists via saveChannelQueries with resolved sort',
-        () async {
-          const filterName = 'sample-app-list';
-          const filterValues = {'user_id': 'test-user-id'};
-          const sortValues = {'pinned_at': true};
-
-          final channelStates = List.generate(
-            3,
-            (i) => ChannelState(channel: ChannelModel(cid: 'test-type-$i:test-id-$i')),
-          );
-
-          when(
-            () => api.channel.queryChannels(
-              predefinedFilter: filterName,
-              filterValues: filterValues,
-              sortValues: sortValues,
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer(
-            (_) async => QueryChannelsResponse()
-              ..channels = channelStates
-              ..predefinedFilter = const PredefinedFilter(
-                name: filterName,
-                filter: Filter.empty(),
-                sort: [SortOption<ChannelState>.desc('last_message_at')],
-              ),
-          );
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer((_) async => <String, List<Message>>{});
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async {});
-          when(
-            () => persistence.saveChannelQueries(
-              cids: any(named: 'cids'),
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              resolvedFilter: any(named: 'resolvedFilter'),
-              resolvedSort: any(named: 'resolvedSort'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              clearQueryCache: any(named: 'clearQueryCache'),
-            ),
-          ).thenAnswer((_) => Future.value());
-
-          await delay(1100);
-          clearInteractions(persistence);
-
-          await client.queryChannelsOnline(
-            predefinedFilter: filterName,
-            filterValues: filterValues,
-            sortValues: sortValues,
-          );
-
-          verify(
-            () => persistence.saveChannelQueries(
-              cids: channelStates.map((s) => s.channel!.cid).toList(),
-              filter: null,
-              sort: null,
-              predefinedFilter: filterName,
-              resolvedFilter: const Filter.empty(),
-              resolvedSort: const [SortOption<ChannelState>.desc('last_message_at')],
-              filterValues: filterValues,
-              sortValues: sortValues,
-              clearQueryCache: true,
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'queryChannelsOffline with predefined filter reads via queryChannelStates',
-        () async {
-          const filterName = 'sample-app-list';
-          const filterValues = {'user_id': 'test-user-id'};
-          const sortValues = {'pinned_at': true};
-
-          final channelStates = List.generate(
-            3,
-            (i) => ChannelState(channel: ChannelModel(cid: 'test-type-$i:test-id-$i')),
-          );
-
-          when(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: filterName,
-              filterValues: filterValues,
-              sortValues: sortValues,
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer((_) async => QueryChannelsResponse()..channels = channelStates);
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer((_) async => <String, List<Message>>{});
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async {});
-
-          await delay(1100);
-          clearInteractions(persistence);
-
-          final channels = await client.queryChannelsOffline(
-            predefinedFilter: filterName,
-            filterValues: filterValues,
-            sortValues: sortValues,
-          );
-
-          expect(channels, hasLength(channelStates.length));
-
-          verify(
-            () => persistence.queryChannelStates(
-              filter: null,
-              sort: null,
-              predefinedFilter: filterName,
-              filterValues: filterValues,
-              sortValues: sortValues,
-              messageLimit: 25,
-              paginationParams: const PaginationParams(),
-            ),
-          ).called(1);
-        },
-      );
-
-      test(
-        'queryChannelsWithResult yields QueryChannelsResult with predefinedFilter=null for inline filter',
-        () async {
-          final channelStates = List.generate(
-            2,
-            (i) => ChannelState(channel: ChannelModel(cid: 'test-type-$i:test-id-$i')),
-          );
-
-          when(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer((_) async => QueryChannelsResponse()..channels = const []);
-
-          when(
-            () => api.channel.queryChannels(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer(
-            (_) async => QueryChannelsResponse()..channels = channelStates,
-          );
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer((_) async => <String, List<Message>>{});
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async {});
-          when(
-            () => persistence.saveChannelQueries(
-              cids: any(named: 'cids'),
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              resolvedFilter: any(named: 'resolvedFilter'),
-              resolvedSort: any(named: 'resolvedSort'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              clearQueryCache: any(named: 'clearQueryCache'),
-            ),
-          ).thenAnswer((_) => Future.value());
-
-          await delay(1100);
-          clearInteractions(persistence);
-
-          final results = await client.queryChannelsWithResult().toList();
-
-          // Persistence returned empty, so only the online emission is yielded.
-          expect(results, hasLength(1));
-          expect(results.single.channels, hasLength(channelStates.length));
-          expect(results.single.predefinedFilter, isNull);
-        },
-      );
-
-      test(
-        'queryChannelsWithResult yields QueryChannelsResult with predefinedFilter populated for predefined query',
-        () async {
-          const filterName = 'sample-app-list';
-          const filterValues = {'user_id': 'test-user-id'};
-          const sortValues = {'pinned_at': true};
-
-          final channelStates = List.generate(
-            2,
-            (i) => ChannelState(channel: ChannelModel(cid: 'test-type-$i:test-id-$i')),
-          );
-
-          const resolvedSort = [SortOption<ChannelState>.desc('last_message_at')];
-          const resolvedFilter = Filter.empty();
-          const expectedPredefinedFilter = PredefinedFilter(
-            name: filterName,
-            filter: resolvedFilter,
-            sort: resolvedSort,
-          );
-
-          when(
-            () => persistence.queryChannelStates(
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer((_) async => QueryChannelsResponse()..channels = const []);
-
-          when(
-            () => api.channel.queryChannels(
-              predefinedFilter: filterName,
-              filterValues: filterValues,
-              sortValues: sortValues,
-              state: any(named: 'state'),
-              watch: any(named: 'watch'),
-              presence: any(named: 'presence'),
-              memberLimit: any(named: 'memberLimit'),
-              messageLimit: any(named: 'messageLimit'),
-              paginationParams: any(named: 'paginationParams'),
-            ),
-          ).thenAnswer(
-            (_) async => QueryChannelsResponse()
-              ..channels = channelStates
-              ..predefinedFilter = expectedPredefinedFilter,
-          );
-
-          when(() => persistence.getChannelThreads(any())).thenAnswer((_) async => <String, List<Message>>{});
-          when(() => persistence.updateChannelState(any())).thenAnswer((_) async {});
-          when(() => persistence.updateChannelThreads(any(), any())).thenAnswer((_) async {});
-          when(
-            () => persistence.saveChannelQueries(
-              cids: any(named: 'cids'),
-              filter: any(named: 'filter'),
-              sort: any(named: 'sort'),
-              predefinedFilter: any(named: 'predefinedFilter'),
-              resolvedFilter: any(named: 'resolvedFilter'),
-              resolvedSort: any(named: 'resolvedSort'),
-              filterValues: any(named: 'filterValues'),
-              sortValues: any(named: 'sortValues'),
-              clearQueryCache: any(named: 'clearQueryCache'),
-            ),
-          ).thenAnswer((_) => Future.value());
-
-          await delay(1100);
-          clearInteractions(persistence);
-
-          final results = await client
-              .queryChannelsWithResult(
-                predefinedFilter: filterName,
-                filterValues: filterValues,
-                sortValues: sortValues,
-              )
-              .toList();
-
-          // Persistence returned empty, so only the online emission is yielded.
-          expect(results, hasLength(1));
-          expect(results.single.channels, hasLength(channelStates.length));
-          expect(results.single.predefinedFilter, isNotNull);
-          expect(results.single.predefinedFilter!.name, equals(filterName));
-          expect(results.single.predefinedFilter!.sort, equals(resolvedSort));
-        },
-      );
-    });
-
-    test('`.disconnectUser` should reset state and user', () async {
-      expect(client.state.currentUser, isNotNull);
-      expect(client.wsConnectionStatus, ConnectionStatus.connected);
-
-      expectLater(
-        // skipping initial connected value
-        client.wsConnectionStatusStream.skip(1),
-        emits(ConnectionStatus.disconnected),
-      );
-
-      await client.disconnectUser(flushChatPersistence: true);
-
-      expect(client.state.currentUser, isNull);
-      expect(client.wsConnectionStatus, ConnectionStatus.disconnected);
-    });
-  });
-
   group('Client with connected user without persistence', () {
     const apiKey = 'test-api-key';
     const userId = 'test-user-id';
